+++ src/Process/Payload/ProcessClientState.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Logistio\Symmetry\Process\Payload;
+
+/**
+ * Interface ProcessClientState
+ * @package Logistio\Symmetry\Process\Payload
+ */
+interface ProcessClientState
+{
+    public function toArray(): array;
+}